+++ essencium-backend-sequence-model/src/main/java/de/frachtwerk/essencium/backend/service/UserService.java
@@ -19,6 +19,7 @@
 
 package de.frachtwerk.essencium.backend.service;
 
+import de.frachtwerk.essencium.backend.configuration.initialization.DefaultRoleInitializer;
 import de.frachtwerk.essencium.backend.model.Role;
 import de.frachtwerk.essencium.backend.model.User;
 import de.frachtwerk.essencium.backend.model.dto.AppUserDto;
@@ -37,8 +38,15 @@ public class UserService extends AbstractUserService<User, Long, AppUserDto> {
       @NotNull PasswordEncoder passwordEncoder,
       @NotNull UserMailService userMailService,
       @NotNull RoleService roleService,
+      @NotNull DefaultRoleInitializer roleInitializer,
       @NotNull JwtTokenService jwtTokenService) {
-    super(userRepository, passwordEncoder, userMailService, roleService, jwtTokenService);
+    super(
+        userRepository,
+        passwordEncoder,
+        userMailService,
+        roleService,
+        roleInitializer,
+        jwtTokenService);
   }
 
   @Override