+++ essencium-backend-uuid-model/src/main/java/de/frachtwerk/essencium/backend/service/UserService.java
@@ -38,8 +38,9 @@ public class UserService extends AbstractUserService<User, UUID, AppUserDto> {
       @NotNull PasswordEncoder passwordEncoder,
       @NotNull UserMailService userMailService,
       @NotNull RoleService roleService,
+      @NotNull DefaultRoleInitializer roleInitializer,
       @NotNull JwtTokenService jwtTokenService) {
-    super(userRepository, passwordEncoder, userMailService, roleService, jwtTokenService);
+    super(userRepository, passwordEncoder, userMailService, roleService, roleInitializer, jwtTokenService);
   }
 
   @Override