+++ essencium-backend/src/test/java/de/frachtwerk/essencium/backend/service/LongUserService.java
@@ -19,6 +19,7 @@
 
 package de.frachtwerk.essencium.backend.service;
 
+import de.frachtwerk.essencium.backend.configuration.initialization.DefaultRoleInitializer;
 import de.frachtwerk.essencium.backend.model.Role;
 import de.frachtwerk.essencium.backend.model.TestLongUser;
 import de.frachtwerk.essencium.backend.model.dto.UserDto;
@@ -35,8 +36,15 @@ public class LongUserService extends AbstractUserService<TestLongUser, Long, Use
       @NotNull PasswordEncoder passwordEncoder,
       @NotNull UserMailService userMailService,
       @NotNull T roleService,
+      @NotNull DefaultRoleInitializer roleInitializer,
       @NotNull JwtTokenService jwtTokenService) {
-    super(userRepository, passwordEncoder, userMailService, roleService, jwtTokenService);
+    super(
+        userRepository,
+        passwordEncoder,
+        userMailService,
+        roleService,
+        roleInitializer,
+        jwtTokenService);
   }
 
   @Override