+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Version `___`
 
+- Feature: An admin user (one of their roles contains all the admin rights defined in the DefaultRoleInitializer) is no longer able to delete themselves.
 - upgraded io.sentry:sentry-spring-boot-starter-jakarta from 7.6.0 to 7.8.0
 
 ## Version `2.5.14`